requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic
python-dotenv
openai
python-multipart
pyjwt
